+++ protocols/gossipsub/CHANGELOG.md
@@ -1,7 +1,12 @@
 ## 0.48.0
 
+- Deprecate `futures-ticker` and use `futures-timer` instead.
+  See [PR 5674](https://github.com/libp2p/rust-libp2p/pull/5674).
 - Apply `max_transmit_size` to the inner message instead of the final payload.
   See [PR 5642](https://github.com/libp2p/rust-libp2p/pull/5642).
+
+- Deprecate `void` crate.
+  See [PR 5676](https://github.com/libp2p/rust-libp2p/pull/5676).
 
 ## 0.47.1