+++ protocols/ping/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.45.1
+
+- Deprecate `void` crate.
+  See [PR 5676](https://github.com/libp2p/rust-libp2p/pull/5676).
+
 ## 0.45.0
 
 <!-- Update to libp2p-swarm v0.45.0 -->